requirements.txt
flask
flask-cors
openslide-python
Pillow
